--- packages/bundler-rpc/src/index.ts
@@ -1,3 +1,0 @@
-export * from './interfaces/IBundler.interface'
-export * from './Bundler'
-export * from './types/Bundler.types'
