+++ packages/bundler/src/utils/httpRequests.ts
